change int to int32 for direct gRPC compability

+++ pkg/graph/permissions.go
@@ -9,7 +9,7 @@ import (
 type Access struct {
 	From driver.DocumentID `json:"_from"`
 	To driver.DocumentID `json:"_to"`
-	Level int8 `json:"level"`
+	Level int32 `json:"level"`
 
 	driver.DocumentMeta
 }
